Tela de departamentos praticamente pronta. Falta apenas a operação de editar, o resto do 'CRUD' já foi feito. Usei também tratamentos de exceção PERSONALIZADO, ficou muito bom e não quebrou a aplicação quando uma exceção foi captada. Consegui também usar a paginação no ListView na tela dos Vendedores. Agora entendi como funciona

+++ WebSales - WebForms e EntityFramework/Services/DepartmentService.cs
@@ -1,9 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
+using System.Data.Entity.Infrastructure;
 using System.Linq;
-using System.Web;
 using WebSales___WebForms_e_EntityFramework.Data;
 using WebSales___WebForms_e_EntityFramework.Models;
+using WebSales___WebForms_e_EntityFramework.Models.Exceptions;
 
 namespace WebSales___WebForms_e_EntityFramework.Services
 {
@@ -33,15 +33,49 @@ namespace WebSales___WebForms_e_EntityFramework.Services
         // Adiciona um departamento
         public void AddDepartment(Department department)
         {
-            _context.Department.Add(department);
-            _context.SaveChanges();
+            // Verifica se já existe um departamento com esse nome
+            if (_context.Department.Where(d => d.Name == department.Name).Any())
+            {
+                // Levantar uma exceção
+                throw new DepartmentException("Já existe um departamento com o nome informado", department.Name);
+            }
+
+            // Tentar adicionar o departamento a tabela
+            try
+            {
+                _context.Department.Add(department);
+                _context.SaveChanges();
+            }
+
+            // Caso dê algum erro de atualização do banco de dados ocorra
+            catch (DbUpdateConcurrencyException e)
+            {
+                throw new DepartmentException("Erro ao adicionar departamento no banco de dados", e.InnerException);
+            }
         }
 
         // Remove um departamento pelo seu Id
         public void RemoveDepartment(int id)
         {
-            _context.Department.Remove(FindDepartment(id));
-            _context.SaveChanges();
-        }       
+            // Verifica se existe algum departamento com esse Id
+            if (FindDepartment(id) == null)
+            {
+                // Levantar uma exceção
+                throw new DepartmentException("O id do departamento informado não existe");
+            }
+
+            // Tentar remover o departamento da tabela
+            try
+            {
+                _context.Department.Remove(FindDepartment(id));
+                _context.SaveChanges();
+            }
+
+            // Caso dê algum erro de atualização do banco de dados ocorra
+            catch (DbUpdateConcurrencyException e)
+            {
+                throw new DepartmentException("Erro ao remover o departamento do banco de dados", e.InnerException);
+            }
+        }
     }
 }
